Exchange: DOUBLE PRECISION to NUMERIC fix

+++ exchange/src/main/resources/db/migration/V1__create_tables.sql
@@ -2,7 +2,7 @@ CREATE TABLE exchange_rate (
     id BIGSERIAL PRIMARY KEY,
     currency_from VARCHAR(10) NOT NULL,
     currency_to VARCHAR(10) NOT NULL,
-    ratio DOUBLE PRECISION NOT NULL DEFAULT 1.0,
+    ratio NUMERIC NOT NULL DEFAULT 1.0,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
@@ -11,9 +11,9 @@ CREATE TABLE exchange_history (
     id BIGSERIAL PRIMARY KEY,
     user_id_from BIGINT NOT NULL,
     currency_from VARCHAR(10) NOT NULL,
-    amount_from DOUBLE PRECISION NOT NULL,
+    amount_from NUMERIC NOT NULL,
     user_id_to BIGINT NOT NULL,
     currency_to VARCHAR(10) NOT NULL,
-    amount_to DOUBLE PRECISION NOT NULL,
+    amount_to NUMERIC NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
